+++ src/main/java/chess/database/LoadGameDao.java
@@ -1,4 +1,4 @@
-package chess.domain.database;
+package chess.database;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
