+++ src/test/java/chess/domain/board/BoardMakerTest.java
@@ -1,10 +1,9 @@
-package chess.domain;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
 
-import chess.domain.board.BoardMaker;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
